feat: 사용자 인증 기반 접근 제어 추가
