+++ frontend/src/pages/_auth.tsx
@@ -1,8 +1,7 @@
 import { Outlet, createFileRoute, redirect } from '@tanstack/react-router';
 
 export const Route = createFileRoute('/_auth')({
-  beforeLoad: ({ context, location }) => {
-    console.log('[context.auth]', context.auth);
+  beforeLoad: async ({ context, location }) => {
     if (!context.auth.isLogin) {
       throw redirect({
         to: '/auth/login',
