+++ frontend/src/shared/api/user/queryKey.tsx
@@ -2,5 +2,5 @@ export const queryKey = {
   all: ['user'],
   signUp: () => [...queryKey.all, 'signUp'],
   login: () => [...queryKey.all, 'login'],
-  users: () => [...queryKey.all, 'users'],
+  user: () => [...queryKey.all, 'user'],
 } as const;
